chore: Refactoring noise method type into primitives.py
